Move `disable` function to new `PushSubscriptionNamespace`
Motivation: namespace will live on `User`
